新增 libs.js，放入 headers 設定
