+++ libs.js
@@ -1,0 +1,12 @@
+const headers = {
+  'Access-Control-Allow-Headers': 'Content-Type, Authorization, Content-Length, X-Requested-With',
+  'Access-Control-Allow-Origin': '*',
+  'Access-Control-Allow-Methods': 'PATCH, POST, GET,OPTIONS,DELETE',
+  'Content-Type': 'application/json'
+}
+
+const libs = {
+  headers
+};
+
+module.exports = libs;
